Migrate `api.test.js`
