+++ tsconfig.json
@@ -13,6 +13,6 @@
     "checkJs": true,
     "noEmit": true,
     "jsx": "react",
-    "types": ["vite/client"]
+    "types": ["vite/client", "@vitest/browser/matchers"]
   }
 }
